+++ mpc_energy/CHANGELOG.md
@@ -6,6 +6,7 @@
 * Increased Amber forecast horizon to use as much as the API returns (can vary between 12-48 hours).
 * Increased MPC forecast horizon to 48hrs
 * Added SolCast day 3 forecast to configuration page (required for 48hr forecast)
+* Improved Streamlit refresh rate to line up with MPC data update rate
 
 ## 0.4.0
 * Added Solar Curtailment Sensor to show when the system is likely curtailing and why